Separated routing-module from app-module for testing purposes
